Refactor problem display to cut down on size
