+++ frontend/src/components/problem/editor/ProblemDisplay.tsx
@@ -1,4 +1,4 @@
-import React, { useCallback, useState } from 'react';
+import React, { useState } from 'react';
 import { useHistory } from 'react-router-dom';
 import { DragDropContext, Droppable, Draggable } from 'react-beautiful-dnd';
 import styled from 'styled-components';
@@ -10,14 +10,14 @@ import {
   problemIOTypeToString,
   ProblemTag,
   TestCase,
-} from '../../api/Problem';
+} from '../../../api/Problem';
 import {
   FixedTextArea,
   PureTextInputTitle,
   CheckboxInput,
   TextInput,
-} from '../core/Input';
-import { Difficulty } from '../../api/Difficulty';
+} from '../../core/Input';
+import { Difficulty } from '../../../api/Difficulty';
 import {
   SmallDifficultyButton,
   PrimaryButton,
@@ -29,23 +29,24 @@ import {
   TextButton,
   InlineLobbyIcon,
   InlineErrorIcon,
-} from '../core/Button';
-import ToggleButton from '../core/ToggleButton';
-import PrimarySelect from '../core/Select';
+} from '../../core/Button';
+import ToggleButton from '../../core/ToggleButton';
+import PrimarySelect from '../../core/Select';
 import {
   SmallHeaderText,
   LowMarginMediumText,
   Text,
   LabelAbsoluteText,
-} from '../core/Text';
-import Loading from '../core/Loading';
-import ErrorMessage from '../core/Error';
-import { FlexBareContainer } from '../core/Container';
-import { generateRandomId, validIdentifier } from '../../util/Utility';
-import { HoverTooltip } from '../core/HoverTooltip';
-import ProblemTags from './ProblemTags';
-import { useAppSelector, useMousePosition, useProblemEditable } from '../../util/Hook';
-import { ProblemHelpModal } from '../core/HelpModal';
+} from '../../core/Text';
+import Loading from '../../core/Loading';
+import ErrorMessage from '../../core/Error';
+import { FlexBareContainer, SettingsContainer } from '../../core/Container';
+import { generateRandomId, validIdentifier } from '../../../util/Utility';
+import { HoverTooltip } from '../../core/HoverTooltip';
+import ProblemTags from '../ProblemTags';
+import { useAppSelector, useMousePosition, useProblemEditable } from '../../../util/Hook';
+import { ProblemHelpModal } from '../../core/HelpModal';
+import OptionsPanel from './OptionsPanel';
 
 const defaultDescription: string = [
   'Replace this line with a short description.',
@@ -66,39 +67,8 @@ const MainContent = styled.div`
   flex: 6;
 `;
 
-const SidebarContent = styled.div`
-  text-align: left;
-  padding: 20px;
-  flex: 4;
-`;
-
 const TopButtonsContainer = styled.div`
   margin-left: auto;
-`;
-
-const SettingsContainer = styled.div`
-  text-align: left;
-  margin: 0.25rem 0 1rem 0;
-  padding: 1rem;
-  border-radius: 10px;
-  box-shadow: 0 -1px 4px rgba(0, 0, 0, 0.12);
-  background: ${({ theme }) => theme.colors.white};
-`;
-
-type ShowProps = {
-  show: boolean,
-};
-
-const ApprovalContainer = styled.div<ShowProps>`
-  display: ${({ show }) => (show ? 'inline-block' : 'none')};
-  text-align: left;
-  margin-top: 0.5rem;
-`;
-
-const ApprovalText = styled(Text)`
-  display: inline-block;
-  margin: 0 0 0 0.75rem;
-  font-size: ${({ theme }) => theme.fontSize.subtitleXMediumLarge};
 `;
 
 const SettingsContainerRelative = styled(SettingsContainer)`
@@ -149,13 +119,7 @@ const StyledMarkdownEditor = styled(MarkdownEditor)`
 `;
 
 const NoMarginTopText = styled(Text)`
-  margin-top: 0px;
-`;
-
-const DeleteButton = styled(PrimaryButton)`
-  margin: 0 0 1rem 0;
-  color: ${({ theme }) => theme.colors.red2};
-  background: ${({ theme }) => theme.colors.white};
+  margin-top: 0;
 `;
 
 const TextButtonLink = styled(TextButton)`
@@ -169,15 +133,6 @@ const TextButtonLink = styled(TextButton)`
 const InlineProblemIcon = styled(InlineLobbyIcon)`
   margin: 10px;
   height: 1rem;
-`;
-
-const InputTypeContainer = styled.div`
-  margin-bottom: 5px;
-`;
-
-const CancelTextButton = styled(TextButton)`
-  margin-left: 2.5px;
-  color: ${({ theme }) => theme.colors.gray};
 `;
 
 type ProblemDisplayParams = {
@@ -201,19 +156,16 @@ function ProblemDisplay(props: ProblemDisplayParams) {
   } = props;
 
   const history = useHistory();
-  const { firebaseUser, token } = useAppSelector((state) => state.account);
+  const { firebaseUser } = useAppSelector((state) => state.account);
   const problemEditable = useProblemEditable(firebaseUser, problem);
 
   const [newProblem, setNewProblem] = useState<Problem>(problem);
   const [loading, setLoading] = useState(false);
   const [error, setError] = useState('');
-  const [hoverVisible, setHoverVisible] = useState<boolean>(false);
   const [helpModal, setHelpModal] = useState<boolean>(false);
 
   // Variable used to force refresh the editor.
   const [refreshEditor, setRefreshEditor] = useState<number>(0);
-
-  const mousePosition = useMousePosition();
 
   const onDragEnd = (result: any) => {
     // dropped outside the list
@@ -230,26 +182,6 @@ function ProblemDisplay(props: ProblemDisplayParams) {
     setNewProblem({ ...newProblem, testCases: newTestCases });
   };
 
-  const deleteProblemFunc = () => {
-    // eslint-disable-next-line no-alert
-    if (!window.confirm('Are you sure you want to delete this problem?')) {
-      return;
-    }
-
-    setLoading(true);
-    setError('');
-
-    deleteProblem(newProblem.problemId, token || '')
-      .then(() => {
-        setLoading(false);
-        history.replace('/');
-      })
-      .catch((err) => {
-        setError(err.message);
-        setLoading(false);
-      });
-  };
-
   // Handle updating of normal text fields
   const handleChange = (e: any) => {
     const { name, value } = e.target;
@@ -259,57 +191,8 @@ function ProblemDisplay(props: ProblemDisplayParams) {
     });
   };
 
-  // Handle updating of enum-type fields
-  const handleEnumChange = (name: string, value: any) => handleChange({ target: { name, value } });
-
   // Handle description change
   const handleDescriptionChange = (value: string) => handleChange({ target: { name: 'description', value } });
-
-  // Handle approval change
-  const handleApprovalChange = (value: boolean) => handleChange({ target: { name: 'approval', value } });
-
-  // Handle updating of problem inputs
-  const handleInputChange = (index: number, name: string, type: ProblemIOType) => {
-    setNewProblem({
-      ...newProblem,
-      problemInputs: newProblem.problemInputs.map((input, i) => {
-        if (index === i) {
-          return { name, type };
-        }
-        return input;
-      }),
-    });
-  };
-
-  const addTag = (newTag: ProblemTag) => {
-    setNewProblem({
-      ...newProblem,
-      problemTags: [...newProblem.problemTags, newTag],
-    });
-  };
-
-  const removeTag = (index: number) => {
-    setNewProblem({
-      ...newProblem,
-      problemTags: newProblem.problemTags.filter((_, i) => index !== i),
-    });
-  };
-
-  // Handle adding a new problem input for this problem
-  const addProblemInput = () => {
-    setNewProblem({
-      ...newProblem,
-      problemInputs: [...newProblem.problemInputs, { name: 'name', type: ProblemIOType.Integer }],
-    });
-  };
-
-  // Handle deleting a problem input for this problem
-  const deleteProblemInput = (index: number) => {
-    setNewProblem({
-      ...newProblem,
-      problemInputs: newProblem.problemInputs.filter((_, i) => index !== i),
-    });
-  };
 
   // Handle updating of test case
   const handleTestCaseChange = (index: number, id: string, input: string,
@@ -355,13 +238,6 @@ function ProblemDisplay(props: ProblemDisplayParams) {
         show={helpModal}
         exitModal={() => setHelpModal(false)}
       />
-      <HoverTooltip
-        visible={hoverVisible}
-        x={mousePosition.x}
-        y={mousePosition.y}
-      >
-        This variable name is likely invalid
-      </HoverTooltip>
       <MainContent>
         <FlexBareContainer>
           <SmallHeaderText>Problem</SmallHeaderText>
@@ -534,132 +410,13 @@ function ProblemDisplay(props: ProblemDisplayParams) {
         {loading ? <Loading /> : null}
         {error ? <ErrorMessage message={error} /> : null}
       </MainContent>
-      <SidebarContent>
-        <SmallHeaderText>Options</SmallHeaderText>
-        <SettingsContainer>
-          <ApprovalContainer
-            show={editMode}
-          >
-            <ToggleButton
-              onChangeFunction={() => handleApprovalChange(!newProblem.approval)}
-              editable={problemEditable}
-              checked={newProblem.approval}
-            />
-            <ApprovalText>
-              {newProblem.approval ? 'Approved' : 'Approval Pending'}
-            </ApprovalText>
-          </ApprovalContainer>
-          <LowMarginMediumText>Difficulty</LowMarginMediumText>
-          {Object.keys(Difficulty).map((key) => {
-            const difficulty = Difficulty[key as keyof typeof Difficulty];
-            if (difficulty !== Difficulty.Random) {
-              return (
-                <SmallDifficultyButton
-                  key={generateRandomId()}
-                  difficulty={difficulty || Difficulty.Random}
-                  onClick={() => (problemEditable ? handleEnumChange('difficulty', difficulty) : '')}
-                  active={difficulty === newProblem.difficulty}
-                  enabled={problemEditable}
-                >
-                  {key}
-                </SmallDifficultyButton>
-              );
-            }
-            return null;
-          })}
-
-          {
-            editMode ? (
-              <ProblemTags
-                problemTags={newProblem.problemTags}
-                addTag={addTag}
-                removeTag={removeTag}
-                viewOnly={!problemEditable}
-              />
-            ) : null
-          }
-
-          <LowMarginMediumText>Parameters</LowMarginMediumText>
-          {newProblem.problemInputs.map((input, index) => (
-            <InputTypeContainer
-              // eslint-disable-next-line react/no-array-index-key
-              key={index}
-            >
-              <TextInput
-                value={input.name}
-                onChange={(e) => handleInputChange(index,
-                  e.target.value, input.type)}
-                disabled={!problemEditable}
-              />
-
-              <InlineErrorIcon
-                show={!validIdentifier(input.name)}
-                onMouseEnter={() => setHoverVisible(true)}
-                // onMouseMove={mouseMoveHandler}
-                onMouseLeave={() => setHoverVisible(false)}
-              >
-                error_outline
-              </InlineErrorIcon>
-
-              <PrimarySelect
-                onChange={(e) => handleInputChange(
-                  index,
-                  input.name,
-                  ProblemIOType[e.target.value as keyof typeof ProblemIOType],
-                )}
-                value={problemIOTypeToString(input.type)}
-                disabled={!problemEditable}
-              >
-                {
-                  Object.keys(ProblemIOType).map((key) => (
-                    <option key={key} value={key}>{key}</option>
-                  ))
-                }
-              </PrimarySelect>
-
-              {problemEditable ? (
-                <CancelTextButton
-                  onClick={() => deleteProblemInput(index)}
-                >
-                  ✕
-                </CancelTextButton>
-              ) : null}
-            </InputTypeContainer>
-          ))}
-
-          {problemEditable ? (
-            <GrayTextButton onClick={addProblemInput}>
-              Add +
-            </GrayTextButton>
-          ) : null}
-
-          <LowMarginMediumText>Return Type</LowMarginMediumText>
-          <PrimarySelect
-            onChange={(e) => handleEnumChange('outputType', ProblemIOType[e.target.value as keyof typeof ProblemIOType])}
-            value={problemIOTypeToString(newProblem.outputType)}
-            disabled={!problemEditable}
-          >
-            {
-              Object.keys(ProblemIOType).map((key) => (
-                <option key={key} value={key}>{key}</option>
-              ))
-            }
-          </PrimarySelect>
-
-          {editMode && problemEditable
-            ? (
-              <>
-                <LowMarginMediumText>Danger Zone</LowMarginMediumText>
-                <DeleteButton
-                  onClick={deleteProblemFunc}
-                >
-                  Delete Problem
-                </DeleteButton>
-              </>
-            )
-            : null}
-        </SettingsContainer>
-      </SidebarContent>
+      <OptionsPanel
+        newProblem={newProblem}
+        setNewProblem={setNewProblem}
+        editMode={editMode}
+        setLoading={setLoading}
+        setError={setError}
+      />
     </>
   );
 }
