+++ frontend/src/views/CreateProblemPage.tsx
@@ -9,7 +9,7 @@ import {
 import { LargeText } from '../components/core/Text';
 import ErrorMessage from '../components/core/Error';
 import Loading from '../components/core/Loading';
-import ProblemDisplay from '../components/problem/ProblemDisplay';
+import ProblemDisplay from '../components/problem/editor/ProblemDisplay';
 import { Difficulty } from '../api/Difficulty';
 import { useAppSelector } from '../util/Hook';
