+++ frontend/src/views/ProblemPage.tsx
@@ -10,7 +10,7 @@ import NotFound from './NotFound';
 import { LargeText } from '../components/core/Text';
 import ErrorMessage from '../components/core/Error';
 import Loading from '../components/core/Loading';
-import ProblemDisplay from '../components/problem/ProblemDisplay';
+import ProblemDisplay from '../components/problem/editor/ProblemDisplay';
 import { generateRandomId } from '../util/Utility';
 import { useAppSelector, useProblemEditable } from '../util/Hook';
